+++ iod/src/rate.cpp
@@ -5,7 +5,7 @@
 static uint64_t microsecs() {
     struct timeval now;
     gettimeofday(&now, 0);
-    return (uint64_t)now.tv_sec * 1000000L + now.tv_usec;
+    return (uint64_t)now.tv_sec * 1000000L + (uint64_t)now.tv_usec;
 }
 
 RateLimiter::RateLimiter() : back_off(boaGeometric), delay(100000), start_delay(100000),
