+++ iod/src/value.cpp
@@ -41,7 +41,7 @@ static bool stringToFloat(const std::string &s, double &x);
 uint64_t microsecs() {
     struct timeval now;
     gettimeofday(&now, 0);
-    return (uint64_t)now.tv_sec * 1000000L + now.tv_usec;
+    return (uint64_t)now.tv_sec * 1000000L + (uint64_t)now.tv_usec;
 }
 
 static const double ZERO_DISTANCE = 1.0E-5;
